Complete feets <-> meters
