+++ src/runtime/providers/string.ts
@@ -30,8 +30,6 @@ class Concat extends Constraint {
   }
 }
 
-
-
 class Split extends Constraint {
   static AttributeMapping = {
     "text": 0,
@@ -99,7 +97,6 @@ class Split extends Constraint {
   }
 }
 
-
 // substring over the field 'text', with the base index being 1, inclusive, 'from' defaulting
 // to the beginning of the string, and 'to' the end
 class Substring extends Constraint {
@@ -148,75 +145,6 @@ class Substring extends Constraint {
   }
 }
 
-class Convert extends Constraint {
-  static AttributeMapping = {
-    "value": 0,
-    "to": 1,
-  }
-  static ReturnMapping = {
-    "converted": 0,
-  }
-
-  resolveProposal(proposal, prefix) {
-    let {args, returns} = this.resolve(prefix);
-    let from = 0;
-    let value = args[0];
-    let to = args[1];
-    let converted;
-    if(to === "number") {
-      converted = +value;
-      if(isNaN(converted)) throw new Error("Unable to deal with NaN in the proposal stage.");
-    } else if(to === "string") {
-      converted = ""+value;
-    } else if(to === "feets") {
-      converted = value * 3.5;
-    }
-    return [converted];
-  }
-
-  test(prefix) {
-    let {args, returns} = this.resolve(prefix);
-    let value = args[0];
-    let to = args[1];
-
-    let converted;
-    if(to === "number") {
-      converted = +value;
-      if(isNaN(converted)) return false;
-      return
-    } else if(to === "string") {
-      converted = ""+value;
-    } else if(to === "feets") {
-      converted = value * 3.5;
-    } else {
-      return false;
-    }
-
-    return converted === returns[0];
-  }
-
-  // 1 if valid, 0 otherwise
-  getProposal(tripleIndex, proposed, prefix) {
-    let proposal = this.proposalObject;
-    let {args} = this.resolve(prefix);
-    let value = args[0];
-    let to = args[1];
-
-    proposal.cardinality = 1;
-    proposal.providing = proposed;
-
-    if(to === "number") {
-      if(isNaN(+value) || value === "") proposal.cardinality = 0;
-    } else if(to === "string") {
-    } else if(to === "feets") {
-    } else {
-      proposal.cardinality = 0;
-    }
-
-    return proposal;
-  }
-}
-
 // Urlencode a string
 class Urlencode extends Constraint {
   static AttributeMapping = {
@@ -259,5 +187,4 @@ class Urlencode extends Constraint {
 providers.provide("concat", Concat);
 providers.provide("split", Split);
 providers.provide("substring", Substring);
-providers.provide("convert", Convert);
 providers.provide("urlencode", Urlencode);
