+++ test/all.ts
@@ -1,3 +1,4 @@
 import "./join";
 import "./eavs";
 import "./math";
+import "./convert";
